Change CirrusSearch sharding values for codfw
During load testing of the codfw cluster it became clear that a
couple servers are serving 2x the query load of others because
they have multiple shards for the same index on them. This patch
adjusts the sharding values for codfw so that we match the number
of primaries + replicas with the number of available servers.

After deploying this the relevant indices need to be rebuilt with
the new shard counts.

Bug: T124215
Change-Id: I26c48589564b5606907d603f42dee3af1a01bf78