Change CirrusSearch sharding values for codfw
During load testing of the codfw cluster it became clear that a
couple servers are serving 2x the query load of others because
they have multiple shards for the same index on them. This patch
adjusts the sharding values for codfw so that we match the number
of primaries + replicas with the number of available servers.

After deploying this the following indices need to be rebuilt with
the new shard counts:
	commonswiki_general, dewiki_content, dewiki_general,
	enwiki_content, enwiki_general, fiwiki_general, frwiki_general,
	itwiki_general, metawiki_general, ruwiki_general

As enwiki takes up 5 times the query volume of the next closest (dewiki)
I can't decided if it's really worthwhile to handle all of these, or
just do enwiki and call it a day?

Bug: T124215
Change-Id: I26c48589564b5606907d603f42dee3af1a01bf78