Have Beta job queue settings shadow production
So iff we're using the job queue on a production wiki, we'll
use it on the Beta equivalent.

Follow up to 8249093a51eb738212f6ae0d509a4cd1abbe99ee .
We are not imminently planning to change production to
use job queue everywhere.

They have already been different a long time, whereas
they only should be for short-term testing in preparation
for the prod change.

Change-Id: Ifa8ea349a5ffde16e407c507575e4974e9ed437d