Delete eqiad masters from codfw configuration and add db weights
codfw configuration should be independent from equiad services.
Pointing to local masters avoids accidental cross-datacenter writes.
(local masters are in read-only mode).

"Dumps,vslow" slaves will not be set to 0, as we do not preview
having load there during the short switchover. A more production
-like config should set them to 0.

Adding weights based on equiad configuration and partitioning
changes that have been rolled back-this needs to be tuned to
maximize both performance and high availability, and for that it
requires stress testing.

Bug: T124697 T124795
Change-Id: I47954e21caf2f63ada0b6e9633025247a8e93902