Prepare db-codfw.php for a live deployment
Add db weights based on eqiad configuration and partitioning
changes-this needs to be tuned to maximize both performance and
high availability, and for that it requires stress testing.

Adding the remote masters has the danger of allowing cross-
datacenter queries, despite the read-only mode. We will keep it
that way until the failover, as per Timo's request.

"Dumps,vslow" slaves will not be set to 0, as we do not preview
having load there during the short switchover. A more production
-like config should set them to 0.

Bug: T124697
Bug: T124795
Change-Id: I47954e21caf2f63ada0b6e9633025247a8e93902