Update patch set 1
Patch Set 1:

This will be needed to go with 1.27.0-wmf.12 which is scheduled to roll to group0 and group1 today. Somebody needs to double check that I'm deferring the config change correctly.

Should fix this HHVM warning now being seen on the beta cluster: "Warning: in_array() expects parameter 2 to be an array or collection in /srv/mediawiki/wmf-config/Wikibase.php on line 120"

Patch-set: 1
Label: Verified=0