Update patch set 4
Patch Set 4: Code-Review-1

-1 pending Aaron's review.

My gut says that MediaWIki db config should never be pointing to local masters as masters as that creates a splitbrain scenario where both are serving different masters instead of the same.

In a switch-over scenario, all DCs will change their perception of what a master is. But in normal operation (and eventually, multi-DC operation) there will still only be one master and we will in fact have read/write queries going to eqiad from codfw in theory. Of course that should be minimised and avoided as much as possible (I think we can realistically reach zero even), by only doing master queries on POST and sending POST requests to the primary DC directly.

However it seems wrong to configure the codfw master as master under normal operation. That creates a fragile situation where code might accidentally start using that (eventhough for both reads and writes it is inaccurate, reads are stale and writes will make it diverge).

Patch-set: 4