Enabling read only mode for s2 before its master failover
The read only is needed for master failover- we must make sure all
changes are replicated and there is no lag before changing the
active master; plus deployment is sadly not transactional.

This has to be complemented with setting the s2-master as read-only
to make sure mediawiki does not ignore this setting in some case.

Bug: T125215
Change-Id: I4d836dabe7a6a59a4a99de9faf5da694f5c12cf6