Monolog: reorder Monolog processors
Processors are added to a logger by MonologSpi in list order using
Logger::pushProcessor() which does an array_unshift(). This means that
the resulting application order of the Processors for a log event is the
reverse of the configuration list order. The 'normalized_message'
processor needs to act on a log record before the 'psr' processor in
order to copy the unexpanded log message. This means that
'normalized_message' needs to come after 'psr' in the configuration
list.

Bug: T124985
Change-Id: I8e7a7f0a6d58987d84ff6255eb97b59a9df49c2a