Replaces "/images" directory with symlinks
Follows-up 145f5cea9f, which moved this dir to /static, but left
a copy for back-compat.

Except for two images (listed below) all files are exactly the same
as those in /static (per md5sum).

Only two different: commons.png and wikimania.png, which were
had their new copies updated in 006b4f2f5f. This commit will
replace those with symlinks to the newer version. The images
in question are of the same type and size and backward-compatible.
(The change just made them transparent for consistency with the
other sul images).

Change-Id: If7a7c4eee8b1d3d230738584a5d1adcd00838718