Clean up old "/images" directory
Follows-up 145f5cea9f, which moved this dir to /static, but left
a copy for back-compat.

As a first step to removing these, this change replaces those images
that are still the same in /static (per md5sum) with a symlink.
This identifies the ones not symlinked as having diverged.

Change-Id: If7a7c4eee8b1d3d230738584a5d1adcd00838718