Update mobile config to use /static/images instead of deprecated /images
As per 145f5cea9f, and the related Varnish configuration that optimizes
paths in /static for all hostnames globally, "/images" is deprecated.

However, MobileFrontend was still using it, and also made updates to it
(20a1e395, f2c67534b, a008bc47). Apply these changes to the "new" copy
instead and replace the old ones with symlinks.

Bug: T107395
Change-Id: I8bdb0ff7632bd2100bf90fe466789dd2d3dbe815