Update patch set 2
Patch Set 2:

It leaves symlinks, so the old-first race condition is fine. The new-first race-condition is valid, but in mw deploys the convention is normally to address that by taking care to deploy it correctly (using separate sync-dir/sync-file operations to handle the order properly) - not by creating separate commits.

Patch-set: 2
Label: Verified=0