Remove live-1.5 symlink to w/ directory
This is an ancient relic from, you guessed it, the 1.5 era of MediaWiki.
It hasn't been used for anything real in basically forever.

I checked a lot of places...puppet, dns, mw-config, MW core, extensions
and nothing I can find still references this path.

There's no need to keep unused symlinks like this around, as it just
makes the directory tree more confusing than it needs to be and
encourages unused things to be used

Change-Id: I0584c1f273c711d9665b41ada1078344a20c8b61