Strip references for experimentation
We will use this change to show the impact of lazy loading references
This should have a more significant impact then navboxes but is unlikely
to tell us anything we don't already know.

Expected impact: 10s shaved off Barack Obama fully loaded time and 50%
of bytes saved on that article with a minor degraation of TTFB.

This patch retains the stripping of navboxes to make this a fair experiment
as otherwise a comparison to the previous 7 days would be an unfair test.

Bug: T126390
Change-Id: I3fe7175c2e24c12f386f5e210ccda7cae687aec1