Speed trials: Add mobile and desktop versions with OOjs UI core loaded
This adds oojs-ui-core to the top script queue, and oojs-ui-core.styles to
the styles queue. It intentionally doesn't do anything with them, as we do
not wish to complicate matters; for now, we are looking at the direct and/
or immediate impact, not the subsequential impacts from e.g. removing from
the bottom queue mediawiki.searchSuggest (for T125725) or the code and the
styling related to watchstars from skins.vector.js and skins.vector.styles
(for T116624).

Bug: T127125
Change-Id: I445440f2c89ee6509d20da88a15f2be2649d5b42