xhgui: sanitize keys
xhgui logs the contents of $_SERVER and $ENV, which can contain sensitive data,
such as the client IP, user-agent string, and cookies. This is then displayed
in the web interface, which means we have to restrict access to it. As a first
step toward making xhgui open-access, send xhgui sanitized copies of $_SERVER
and $ENV, which contain only explicitly whitelisted keys.

Change-Id: I219c090110871c16b576a4d6b5e16825a9872ff1