Update patch set 1
Patch Set 1:

Given that es2001 has a different hardware from the new es201*, I thought that if possible would be better to test machines with same hardware.

One option is to have one of the new es201* reformatted with a stripe of 256 KB (and maybe one with 1 MB) and compare that with es2011 performances.

Another option could be to use one of the es1 in eqiad, looks like the same hardware and the load is low so should not be a problem to have one depooled for a bit.

Patch-set: 1