wmfstatic: Allow longer client-side caching of 'nohash' responses
Currently nohash is treated as a penalty way (only 5 minutes).
In reality, while it's nice to strive for having all urls hash-validated,
at the moment this isn't feasible because there are many circumstances in
which we can't (yet) produce or transport a hash to the code making the url.

Also, there are also valid use cases for explicitly not wanting to pin a
url to a specific verison fo a file when that would cause different pages
to display different version of the file (due to static caching of html
output,e.g. the "Powered by" logo in the footer and other interface elements
not provided by a central stylesheet).

An exception to the above is the site logo because that isn't linked from HTML.
It's referenced by a stylesheet that itself is centrally loaded and invalidated
(so it never varies by page).

For in-HTML images (such as the footer images and other interface images), one
doesn't want to hardcode the file version.

Recognise this by bumping the allowed cache age for clients from 5min to 24h.
I'm keeping the server-side max-age at 5 min for now to allow allow deployments
to quickly affect users with no prior cache. It's cheap enough to have the
backends bounce a 304 Not Modified around every few minutes for hot file paths
(similar to how it does now).

While this does increase max-age for some urls, it effectively still shortens
it compared to how it was before wmfstatic existed. Previously these urls were
always cached for a year unconditonally. And the majority of wikis isn't using
wmfstatic yet.

Change-Id: Iee2f359cf474347b03360e1da97c605c16aebbc6