Repool db1021 and db1024, both with low/non critical load
Those are a host which suffered corruption and has since been
rebuilt from the master, and the previous master, that was
defragmented. They have been offline for a while, and due to
previous hardware issues I am being conservative (although they
have been up and without issues for weeks already).

Bug: T126996 T120122
Change-Id: I039ea6d6ba7da3524a9b01dfc4810c325b1fbd7c