Update patch set 4
Patch Set 4:

Luke081515, the last time any of these groups were renamed was in Aug 2014, as part of an effort to standardize the group names. If a group name changes, there a bunch of pages/messages that also have to be updated. Ideally I agree, password policies shouldn't get removed because of a renaming, but unfortunately CentralAuth tracks everything else about groups by name.

Patch-set: 4