Update patch set 1
Patch Set 1:

Looks good to me, and most of all not dangerous. We will have to review the base value for pool counter once we enable SSL (no idea what the overhead is, but probably significant, and not the same inside / across DC). This should probably done in another change.

Patch-set: 1