Update patch set 1
Patch Set 1: Code-Review+1

Try to warm up the buffer pool (as they are not replicating, they will not do it naturally)-not a blocker for repooling.

What I did last time was to compute some counts(*) using the primary key- which at the same time allowed me get some heuristics to check with the old servers that the copy was done correctly (to be fair, it is extremely difficult to not do it right, but one never knows if we can discover a mediawiki bug, etc.).

Patch-set: 1
Label: Code-Review=+1