Update patch set 7
Patch Set 7:

This is ready for review, please complain about any potential syntax or logical error, but let me deploy as soon as possible to perform proper load testing (as this is not yet live).

It will probably not be the final config, but it is the best I can do as a guesstimate, based on eqiad experience. I will be able to tune roles and load better after testing.

Patch-set: 7