Stop pushing ES updates to nobelium
We've completed our testing to measure the ability of elasticsearch
to keep up with indexing volume on spinning disks. The results are not
amazing, but it gave us the info we needed. Turn off writes to open
up nobelium's IO for use measuring querying indices which don't fit in
memory rather than write performance.

Change-Id: I38308fc68e58486b98ccc82361eadb57943552c0