Update patch set 12
Patch Set 12:

@Jcrespo	: I'm not doubting that authority in the least. Let me clarify.

I had good reasons to assume that it is possible to cause incorrect application logic when assigning a "fake" master as master. E.g. bad reads from "master" populating secondary data stores. As well as incompatibilities with the fact that our current MediaWiki deployment cannot fully operate error-free without being able to connect to the "real" master (read-only mode is not the same as masterless mode, which is what codfw would be)

E.g. GET requests try to establish master connections due to technical debt in various extensions that we are actively trying to get rid of, but that takes time. At least Flow and Translate extensions both have or had recently still models in place that logically depend being able to read data from the true master - which we all now is bad and we're working on phasing that out as quickly as we can 

I have indeed previously expressed that using eqiad-master for codfw is preferred as long as eqiad is primary. However that is merely a preference because I know that to work properly (from an application logic perspective).

It's okay if it's unacceptable from a DB perspective to have Codfw talk to Eqiad master (even if such connection would be rare). I don't mean to change or doubt that.

I'm merely saying that if we must go that way, we will of course have to make sure that MediaWiki actually supports that. And make the adequate changes in the software. We are responsible for preventing data corruption and ensuring application works as expected (uptime isn't very relevant if only to corrupt user contributions). As such, worst case scenario we'll have to block use of codfw app servers until those guarantees are put in place within MediaWiki.

I have no interest in preferring codfw's use of eqiad-master beyond ensuring that MediaWiki works properly :)

Patch-set: 12