Revert "(bug 45233) Groups permissions on pt.wikivoyage"
The pt.wikivoyage community requested in 2013 the possibility
for sysops to grant bot flag. This right has been used in 2013,
but not for three years. The community is now less active than
at the request.

Meanwhile, stewards expressed concerned we're unweaving the bureaucrat
status, and small communities lack a critical size to exercise check
and balances on a rather powerful right.

It so seems more careful to revoke this possibility, as it's not used
anymore and isn't supported by the stewards.

This reverts commit 9da27de8d85cf93ba4e983a707bc27ddbb9e21fd.

Bug: T129487
Change-Id: I3723019baf2fb531f743d5dada3d649be763ff29