Use 'include' instead of 'include_once' in missing.php
Follows-up c1200c3.

The 'once' instruction doesn't make sense here since the interwiki map
is accessed by return value. So if other code did load it before, we need
to load it again here.

If the interwiki map was a constant or class, then we'd require it once
and access the exported identifier but in this case we need the file to
be loaded regardless.

When include_once is called a second time for the same file, it returns
boolean true instead of the file's return value.

Change-Id: I209ece8eea128057cf03d789db8a6073cb99f009