Revert "Remove extraneous namespace shorcut alias on ru.wikibooks"
Change Ie6ded5f53f038b7c6cd7af3c4d9ff3a5f5410729 correctly introduced
a shorcut namespace alias М (cyrillic letter) for Module.

During the test, it were confused with a latin M, and so discarded,
as conflicting with the M for meta.

It appears the error was the test procedure, not the change.

This reverts commit ec9f9aca9a446f7e14c20824171f2745e8d46152.

Change-Id: I9d2f97832889a3cd1c1adfa8315a549c4c32dc83