Update patch set 2
Patch Set 2:

There is tooling support for XWD:1 (a Firefox plugin and a Chrome plugin) but not for XWD:log, IMO that should be fixed first.

Are we sure requests with XWD header always go to one of the designated machines? Even requests to upload.wikimedia.org or those which will be routed to static.php?

Patch-set: 2
Reviewer: Gergő Tisza <389@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0