Update patch set 4
Patch Set 4:

Note it may be more worthwhile to actually deploy opus audio in .webm container instead of Ogg/.opus, as Matroska/WebM container is being implemented for MS Edge future builds while ogg may or may not. That'll probably need a tweak to the code, I know we support those on input but not sure as a transcode target.

Patch-set: 4