Update patch set 2
Patch Set 2:

I am not worried about that, that is supposed to be fixed. I was giving my ok to deploy it as the infrastructure side of things has worked properly for weeks now.

My concern is that during a short period of time (the actual master failover; even in read-only mode):

1) either we will have 2 instances of pt-heartbeat running
2) no instance will be running

Which in theory, that is why I encourage you to do a group by + max - to allow for overlapping, but we never know.

There are other things, but those are related to within-datacenter master-failovers and active-active, so out of scope.

Patch-set: 2