Remove T44894 FIXME note
The Swahili community has decided to restrict page creation for IPs.

Plan was to reevaluate the situation around 2015-09-15. It has then
be decided to keep the restriction, as there aren't any volunteers
on sw.wikipedia comfortable to use AbuseFilter and implement filters.

We can so discard the comment to revisit the issue.

Change-Id: I49a0169b73809702196c7d4eeec90e13286b1136