Configure a single elasticsearch server for CirrusSearch beta cluster
As we are doing client side load balancing in beta, we do not reuse
connections (new HTTP calls are likely to go to a different server
than the one last used by our curl handle).

This is a temporary change that will need to be reverted after testing is
done.

Bug: T130219
Change-Id: I6fec85b3392f36984bca478d668a1658c21af8db