Update site association file to support handoff and credential sharing with the iOS app
This patch adds 2 additional dictionaries to
the site association file:

activitycontinuation:
This enables Handoff between the Safari
browser on any Mac/iOS device and
the Wikipedia iOS app.

See here for more information:
https://developer.apple.com/library/ios/documentation/UserExperience/Conceptual/Handoff/HandoffFundamentals/HandoffFundamentals.html

webcredentials:
This enables sharing of credentials for
Safari and the iOS app in the iOS keychain.
This means 2 things:
1. If a user logs in to the app, then
those credentials will be available
to login in Safari.
2. If a user logs into Wikipedia.org,
then those credentials will be available to
login in the iOS app.

See here for more information:
https://developer.apple.com/library/ios/documentation/Security/Reference/SharedWebCredentialsRef/

Change-Id: Ia2e83cc3e268cd32f467a8e3d5c264efdad77919