Update patch set 3
Patch Set 3: Code-Review-1

[ JSON format ]

JSON would be more readable if you keep the non minimized version. If there is a need to minimize the file for performance issue, we need a non minimized version to maintain it.

[ Commit message ]

First line of the commit message should be max 72 characters, so it's not mangled on Gerrit, Phabricator Diffusion or GitHub. Ideally, keep it < 50 characters to help humans.

To get an association between task and change in our current CI infrastructure, instead of a T128795 lost at the top, add a "Bug: T128795" just before Change-Id.

If you want to underline, please use a fixed width font, ie use the same number of characters in your heading and the dashes line.

Patch-set: 3
Label: Code-Review=-1