Update patch set 7
Patch Set 7:

For the configuration, there isn't really any "releases" schedule: config changes are deployed one per one, or by group of similar changes, either during SWAT for low risk, rather trivial requests, either during specific deployment windows for more risky things or new features, either on the fly when there is an emergency or it's a regular part of operations.

So for this kind of changes, yes, SWAT is the normal process.

Patch-set: 7