CirrusSearch: Add new rescore profiles
In order to run optimization plans we need these rescore profiles enabled in
production.
This should be a temporary solution, we will run various optimization with
large samples to evaluate if any of these new functions are worth the effort.
These profiles are also suited for A/B tests: all the values can be overridden
by a request param.

I'm not sure that wmf-config is the best place to store these profiles. They
are particularly verbose and will certainly require fine-tuning per wiki. I
wonder if a datastore (RESTBase?) might be more appropriate, this would allow
us to do live optimizations without deploying anything...

NOTE: depends on 1.27.0-wmf.19

Bug: T127896
Change-Id: I17aab1177662a5dad2c513ad1b406f3fd155ecc4