Revert "Correct HTML code for WMF image"
This was the result of a misunderstanding. There is nothing incorrect
about it. We emit quoteless attributes in MediaWiki by default where
legal by the HTML syntax. This is a relatively unimportant optimisation
but there is no point in explicitly adding quotes where previously
generated content omitted them. They'll just be removed again next time
such a piece of content is generated.

This reverts commit aca5c55c293e0eef79fdb2ee890f1696bab3f99d.

Change-Id: I356544a5be9689b893598c971d3550fb3b1275ce