Match JsonConfig change
This patch needs to be deployed to match
the change in how JsonConfig parses its configuration.

Deploying it now is a noop because 'pattern' parameter
does not yet exist.

https://gerrit.wikimedia.org/r/#/c/281327/
change I572299ddad8c5e6366d5eb0b9d708138c9fc1b19

Now there is a 'pattern' parameter which decides
which page in the given namespace should be handled.

Change-Id: Iec0421ecce668747f868cf02942ab7137c9e6c05