Update patch set 3
Patch Set 3: Code-Review-1

We should not do this without a clearer understanding of how it is going to play together with the queries we have planned for on-wiki. I'd like us not to get into a situation where we give people several ways to do the same thing without a good reason for that. That just creates a lot of confusion.

Patch-set: 3