Update patch set 1
Patch Set 1: Code-Review-2

We need to be able to document map usage for Wikivoyage, and therefor show how it works. MediaWiki.org is used explicitly for help pages, and therefor it should be possible to have all capabilities enabled there. Per our agreement with Ed, we will not enable this capability on Wikipedia, unless we reach an agreement on that in a separate discussion.

Patch-set: 1
Label: Code-Review=-2