Reconfigure interface editor group on ur.wikipedia
Previously, per T120348, sysops were able to grant and revoke
the right to edit interface.

As the interface edit right requires an high level of trust,
it seems there is a consensus to restrict this possibility to
bureaucrats.

Accordingly, we reconfigure the interface editor group to be
managed by bureaucrats.

This is a follow-up for f038ac38.

Bug: T133564
Change-Id: I45b53d7e758a31428a2568262aeb1df5524adbc7