Update patch set 1
Patch Set 1: Code-Review-1

i. This is not only a translation, but a new namespace request, and so need to be discussed at https://diq.wikipedia.org/wiki/Wikipedia:Portal%C3%AA_cemaeti or any other appropriate place on the project. As a very small community, a notification without objection after some days is enough, a formal vote isn't required. But in all case, we want such on-project discussion to offer the opportunity to the community to give feedback.

ii. Technically, this should be manually rebased to use short array syntax introduced in 8d4cf82c.

Patch-set: 1