Enable lazy loaded references in beta
This enables lazy loaded references in enwiki to allow us to get a sense
of how this change may impact global performance and whether it is worth pursuing further.

This will temporarily break references for non-JS users in the beta channel but given traffic there is low and this may save us lots of work in the long term this is deemed acceptable.

Bug: T129693
Change-Id: I292c467f14dbeb452d9a3001c45f8c7227b32649