Set $wgRateLimits['upload'] for Commons
This was previously implemented using an abuse filter, which doesn't
work very well with various tools. This configuration is meant to
match the existing one exactly.

This is an exact copy of the rules in filter 140. The groups listed
are all that have the 'autopatrol' right but don't have 'noratelimit'.

Bug: T132930
Change-Id: I28730d5bac4f83f9ef5950a2babc6d45d119b143