Update patch set 1
Patch Set 1:

So, aside from James's suggestion that /testwiki2?/ is a terrible thing, I'm curious if the multiple foreign repository thing has been tested when looking at, e.g., different files on different sites with the same filename. If not, maybe we should consider looking at that before we deploy this, unless we don't care about testwiki.

Oh, wait, we don't.

OK this is probably fine, but as an academic exercise I'd be interested to know if multiple FFRs will break in this way.

Patch-set: 1
Label: Verified=0