Remove redundant NS_PROJECT entries from wgNamespacesAliases
Each family has the English project name as NS_PROJECT alias for all
the dblist. We so prune individual entries.

We also prune a redundant WV alias for de.wikivoyage, as each wiki
in wikivoyage dblist has an WV alias for NS_PROJECT.

Thanks to @MarcoAurelio to have noticed this pattern.

Bug: T131023
Change-Id: I9eaf26f47c8a6b69de7c095463b08942ca8edb20