Update patch set 4
Patch Set 4:

No, the patch was fine. It was a temporary deployment issue because I used sync-dir wmf-config/ instead of doing sync-file InitialiseSettings before CommonSettings. And it looks like there are still issues because the messages get buffered on the app servers, but they would have stopped after a few seconds.

Patch-set: 4