Update patch set 4
Patch Set 4:

Well your patch is entirely legit. Legoktm approved it and I would have approved it as well.

As for why the variable is not set at this point, that is a mystery. Legoktm is looking at it. Feel free to join the IRC #wikimedia-operations channel :) We are both in there.

Patch-set: 4