Set Wikibase repoUrl to use https
ArticlePlaceholder uses repoUrl to build the url for Wikidata's
api.

If it is protocol relative, then I think it defaults in
this case to http and then the request might be redirected normally
to https, but not in this case and we get an empty response.

This causes ArticlePlaceholder search result suggestions to
not appear on Special:Search when they are supposed to.

PS. Ideally we do the item lookup another way, internally,
with EntityLookup or lookup the relevant page props on Wikidata
to get the info that ArticlePlaceholder needs.

Change-Id: I7b1ac171e8733a225a29a0ae34085f06dd296d9e