Set $wgDisableAuthManager = true
This is a no-op that allows the default value to be set to false
in core without affecting the production cluster. Core and all
bundled extensions have been updated for authmanager but not all
extensions used by WMF have, so it's OK to enable by default
but it has to stay disabled in production a while longer.

Bug: T135498
Change-Id: Iaa47e8ce97081f40c081c02567e9b2b80d39f8db