Update patch set 1
Patch Set 1:

It has to be a file hash so that it can be verified. Among other thigns this enables the fundamental part of multiversion which is to be able to access a file through a domain-relative url regardless of the deployed wiki version.

For example, /w/extensions/ULS/foo.woff2?12345 would make sure that version 12345 is served, regardless of whether the wiki just got rollback, or upgraded, or in the middle of a deployment etc. And even just under normal operation it will default to the version related to the current wiki hostname. But because that can change (and we don't want them to get stuck in Varnish for a week without being able to deploy updates to it), the url must change and be verifiable.

If the url only changes with deployments, it'll solve the cache-bypass problem, but not the verification problem. It happens quite often that a url changes (consumer got HTML or JS from server 1) and then bypasses Varnish and hits server 2 (which is still waiting for scap) and thus populates the old content again at the new url. This won't happen here since it is verified and thus cache age is shortened to 5 minutes in case of a mis-match so that it naturally resolves itself. (During this, it falls back to the newest version of the file that exists in any of the deployed branches).

Patch-set: 1