Update patch set 4
Patch Set 4:

Remeber that userrights-interwiki makes only real sense together with userrights, because userrights-interwiki only gives you the permission to manage rights interwiki, but not which rights. So to manage all, you need userrights too. (If it's assigned via a global group to the users of this group it is ok, but if not, we should add it)

Patch-set: 4