Update patch set 3
Patch Set 3:

Yes, it was.

As said before, your next responsibility is to ensure this change is deployed, and test if it works as expected during the deployment.

To do so, you can add this patch to one of our deployment windows.

See https://wikitech.wikimedia.org/wiki/SWAT_deploys for more information.

We can merge and deploy it 8 windows per week, Monday to Thursday, at 17:00 or 1:00 CEST.

Pick the one the more convenient for you and add this Gerrit change to https://wikitech.wikimedia.org/wiki/Deployments:

{{ircnick|Local Profil or your IRC nick}}
* [config] {{Gerrit|288582}} Make SUL icons square and use global defaults ({{phabT|135212}})

Patch-set: 3