Send wmf.4 search and ttmserver traffic to codfw
wmf.3 contains search code for talking to elasticsearch 1.7, wmf.4
contains code for talking to 2.3. As the train rolls forward and wiki's
switch the code they are running they also need to switch which
elasticsearch cluster they are talking to.  Patch adds a temporary hack
which points all wmf.4 traffic at codfw which will be running 2.3.

Also removes the hard coded wgCirrusSearchMoreLikeThisCluster, allowing
it to fall back to whatever the default search cluster is.

Bug: T133124
Change-Id: I9e2dedb80d59fec8d13f0eb67e8980318ef3d130