Make VE RB URLs domain-relative
So that on mobile you go to the mobile domain for RB, otherwise the normal
domain.

This is to help clients that don't honor RB's Access-Control-Expose-Headers
cross-origin.

As a side-effect, API requests from mobile users should be sped up slightly by
reusing the existing connection, and avoiding the extra DNS lookup.

Bug: T135171
Change-Id: I825b22058e7e1c536298a4a19e0ff1981045a360