Update patch set 1
Patch Set 1:

The documentation around enabling an experiment does not seem to be very helpful. I'm entirely confused about what the name of the property should be in InitialiseSettings.php, if there are supposed to be multiple properties (wmgPopupsExperiment and wmgPopupsExperimentConfig?) and how to properly set them per wiki. I'll upload a new patch set with how I think it works, but the documentation differences between mw.experiments and Popups/extension.json are very confusing.

Patch-set: 1
Label: Verified=0