Update patch set 1
Patch Set 1:

That's the ores extension setting for beta wikis. For now, we are checking to see if the prod environment can work with the ores extension and nothing wrong with it. Then, we move to use the prod service in production wikis. i.e. the ORES in prod is in testing period atm.

Patch-set: 1
Reviewer: Ladsgroup <920@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0