Add tags for group1 and group2
We ran into issues when trying to deploy AuthManager to group0 and
group1 wikis: Setting enabled for "group1" didn't work since that tag
doesn't exist, and setting the default enabled and disabling for
"wikipedia" (as a proxy for the missing "group2") didn't quite work
right either since testwiki and test2wiki are in that group. To avoid
similar confusion in the future, let's just make "group1" and "group2"
be valid configuration tags.

This also adds a group2.dblist, so it can be loaded to define the tag.

Change-Id: I8ff64b8b497dcf354e6bff4d4d81a5e13678df0b