Update patch set 1
Patch Set 1: Code-Review-1

Very weak community consensus. Sure the request proponent confirms it's okay to proceed, but (i) there is an opinion from a contributor this is a weak consensus for ko. (ii) an established technical contributor  has chosen not to comment the issue.

We should perhaps ask the original author to gather more support on the local wiki.

Patch-set: 1