Complete list of legacy main pages, switch default to false
After auditing which wiki projects use the legacy wgMFSpecialCaseMainPage
option, add them all to the mobilemainpagelegacy dblist.

This will force all new wikis and existing wikis not currently using
the special casing to be forced to not use it and educate existing
projects to move away from using it.

Additional changes:
* Add wikispecies to the all.dblist

Bug: T138425
Change-Id: I0d5e658663fef08bd1cae1d059789af6b8145bee