English Wikipedia uses legacy main page formatting
English Wikipedia uses a table based layout.
It should have been in the mobilemainpagelegacy.dblist
but was excluded from the results of the audit scripts
due to the fact it uses
mp- prefixes rather than mf- prefixes (this is badly
documented but apparently supported)

I am going to re-run the audit script to check for mp- prefixes
in case other wikis are impacted. I'll fold them into this patch
before SWAT.

Bug: T138578
Change-Id: Ib6e448b27a9dd7635d325e8f7b0ca9dac32ae5af