Update logo settings for the Nepali Wikipedia
This patch updates the logo for the Nepali Wikipedia
(newiki) and adds the usual high-density versions.

The creation of this patch was discussed by myself
and Krish Dulal, a member of the Nepali Wikimedians
of Nepal user group, during the 2016 Wikimania.

The logo itself has already been updated on Commons
by me and this patch follows that change.

Bug: T139240
Change-Id: I70f0a8224c7fbfe1ecf5487978344454d908e0bf